requirements.txt
Django
django-ckeditor
django-cleanup
django-filter
django-jazzmin
django-phonenumber-field
django-cors-headers
djangorestframework
django-modeltranslation
django-rest-registration
django-resized
drf-writable-nested
drf-yasg
pandas
phonenumbers
Pillow
psycopg2-binary
python-decouple